refactor(config): move validation utilities to separate file
- Extracted ZogUInt, ZogUInt64 and related validation functions from config.go
- Created new validators.go file to contain all validation utilities
- Simplified validation function implementations by extracting common coercer logic
